+++ jericho/models.py
@@ -66,6 +66,7 @@ class JerichoDnsServers(Base):
     time_created = Column(DateTime(timezone=True), server_default=func.now())
     id = Column(Integer, autoincrement=True, primary_key=True)
 
+
 class JerichoConverter(Base):
     __tablename__ = "jericho_converter"
 
@@ -73,6 +74,7 @@ class JerichoConverter(Base):
     location = Column(String(255))
     time_created = Column(DateTime(timezone=True), server_default=func.now())
     id = Column(Integer, autoincrement=True, primary_key=True)
+
 
 class JerichoWorkload(Base):
     __tablename__ = "jericho_workload"